Refactoring of the input utils into pluggable classes
