+++ dask_sql/input_utils/base.py
@@ -1,0 +1,11 @@
+from typing import Any
+
+
+class BaseInputPlugin:
+    def is_correct_input(
+        self, input_item: Any, table_name: str, format: str = None, **kwargs
+    ):
+        raise NotImplementedError
+
+    def to_dc(self, input_item: Any, table_name: str, format: str = None, **kwargs):
+        raise NotImplementedError
